Extract subscription hook from UI component
